add the functionality to prevent the user from deleting notes that do not belong to them.

+++ src/note/note.service.ts
@@ -1,4 +1,9 @@
-import { BadRequestException, Body, Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Body,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { Prisma } from '@prisma/client';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateNoteDTO } from './dto/create-note.dto';
@@ -28,9 +33,11 @@ export class NoteService {
     const { authorNickname } = data;
 
     if (!data.title || !data.content || !data.authorNickname) {
-      throw new BadRequestException('Preencha todos os campos para criar uma nota.');
+      throw new BadRequestException(
+        'Preencha todos os campos para criar uma nota.',
+      );
     }
-        
+
     await this.existsAuthor(authorNickname);
 
     const noteData: Prisma.NoteCreateInput = {
@@ -60,14 +67,24 @@ export class NoteService {
     await this.existsNote(id);
     return this.prisma.note.update({
       data,
-      where: { id }
+      where: { id },
     });
   }
 
-  async deleteNote(id: number) {
+  async deleteNote(id: number, userNickname: string) {
     await this.existsNote(id);
+
+    const note = await this.prisma.note.findUnique({
+      where: { id },
+      select: { authorNickname: true },
+    });
+
+    if (note.authorNickname !== userNickname) {
+      throw new BadRequestException('Você não tem permissão para excluir a nota de outro usuário.');
+    }
+
     return this.prisma.note.delete({
-      where: { id }
+      where: { id },
     });
   }
 }
